+++ js/utils.js
@@ -1,0 +1,57 @@
+const generateArray = (length, max) => (
+  [...new Array(length)]
+    .map(() => Math.round(Math.random() * max))
+);
+
+const checkPositiveArray = (initialNumber, finalNumber) => (initialNumber < 0 || finalNumber < 0) ? 0 : 1;
+
+const swapValues = (initialNumber, finalNumber) => {
+  if (finalNumber < initialNumber) {
+    finalNumber = [initialNumber, initialNumber = finalNumber][0];
+  }
+
+  return [initialNumber, finalNumber];
+};
+
+const getRandomPositiveInteger = (initialNumber, finalNumber) => {
+  if (checkPositiveArray(initialNumber, finalNumber) === 0) {
+    return;
+  }
+
+  [initialNumber, finalNumber] = swapValues(initialNumber, finalNumber);
+
+  initialNumber = Math.ceil(initialNumber);
+  finalNumber = Math.floor(finalNumber);
+
+  return Math.floor(Math.random() * (finalNumber - initialNumber + 1)) + initialNumber;
+};
+
+const getRandomPositiveFloat = (initialNumber, finalNumber, presicion = 5) => {
+  if (checkPositiveArray(initialNumber, finalNumber) === 0) {
+    return;
+  }
+
+  [initialNumber, finalNumber] = swapValues(initialNumber, finalNumber);
+
+  return (Math.random() * (finalNumber - initialNumber) + initialNumber).toFixed(presicion);
+};
+
+const createElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const createRandomArray = (elements) => {
+  const lengthArray = getRandomPositiveInteger(1, elements.length);
+  const newArray = [];
+  do {
+    const newElement = createElement(elements);
+    if (!(newArray.includes(newElement))) {
+      newArray.push(newElement);
+    }
+  } while (newArray.length < lengthArray);
+  return newArray;
+};
+
+export {generateArray};
+export {getRandomPositiveFloat};
+export {getRandomPositiveInteger};
+export {createRandomArray};
+export {createElement};
